+++ lib/json_matchers.rb
@@ -1,3 +1,4 @@
+require "pathname"
 require "json_matchers/version"
 require "json_matchers/matcher"
 require "json_matchers/errors"
@@ -8,6 +9,6 @@ module JsonMatchers
   end
 
   def self.path_to_schema(schema_name)
-    Pathname(schema_root).join("#{schema_name}.json")
+    Pathname.new(schema_root).join("#{schema_name}.json")
   end
 end